requirements.txt
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
distconfig3==1.0.1
exceptiongroup==1.1.2
graphql-core==3.2.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pydantic==2.1.1
pydantic_core==2.4.0
pytest==7.3.0
PyYAML==6.0.1
requests==2.31.0
sgqlc==16.3
structlog==23.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
vyper-config==1.1.1
watchdog==3.0.0
pyhamcrest
allure-pytest